item model moved to models and model folder deleted
