+++ webapp/Components/HowItWorksComponent.razor
@@ -1,0 +1,44 @@
+@inherits LayoutComponentBase
+
+<section class="how-works section" style="background-color: #fff;">
+    <div class="container">
+        <div class="row">
+            <div class="col-12">
+                <div class="section-title">
+                    <h2 class="wow fadeInUp" data-wow-delay=".4s">How it Works</h2>
+                    <p class="wow fadeInUp" data-wow-delay=".6s">There are many variations of passages of Lorem
+                        Ipsum available, but the majority have suffered alteration in some form.</p>
+                </div>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-lg-4 col-md-4 col-12">
+                <!-- Start Single Work -->
+                <div class="single-work wow fadeInUp" data-wow-delay=".2s">
+                    <span class="serial">01</span>
+                    <h3>Create Account</h3>
+                    <p>Lorem ipsum dolor sit amet constur adipisicing sed do eiusmod tempor incididunt labore.</p>
+                </div>
+                <!-- End Single Work -->
+            </div>
+            <div class="col-lg-4 col-md-4 col-12">
+                <!-- Start Single Work -->
+                <div class="single-work wow fadeInUp" data-wow-delay=".4s">
+                    <span class="serial">02</span>
+                    <h3>Post Your Ads</h3>
+                    <p>Lorem ipsum dolor sit amet constur adipisicing sed do eiusmod tempor incididunt labore.</p>
+                </div>
+                <!-- End Single Work -->
+            </div>
+            <div class="col-lg-4 col-md-4 col-12">
+                <!-- Start Single Work -->
+                <div class="single-work wow fadeInUp" data-wow-delay=".6s">
+                    <span class="serial">03</span>
+                    <h3>Sell Your Item</h3>
+                    <p>Lorem ipsum dolor sit amet constur adipisicing sed do eiusmod tempor incididunt labore.</p>
+                </div>
+                <!-- End Single Work -->
+            </div>
+        </div>
+    </div>
+</section>
